+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 - Add logging for terminal and discover
+- Default friendly name if user doesn't provide one (TSP-757)
 
 ## [0.16.4]